Move yaml config specific schemas into separate file
